# One-shot decrypt directly in to destination with padding