@vcsjones 2022-02-09
@bartonjs I pushed up some changes based on your feedback. New benchmarks below.

tl;dr: the large data sizes increased as one would expect due to the less copying. Small data regressed, but, I shortcut those to do a single call in a stack buffer which makes small data (<=128) "happy" again.


<details>

|                           Method |        Job | Toolchain | DataSize | Padding |         Mean |       Error |      StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------------- |----------- |---------- |--------- |-------- |-------------:|------------:|------------:|------:|--------:|-------:|------:|------:|----------:|
| Decrypt_Cbc_GuaranteedBufferSize | Job-SSHGXV |    branch |       16 |   PKCS7 |     580.2 ns |     6.74 ns |     6.30 ns |  0.96 |    0.01 | 0.0200 |     - |     - |     128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZZAIDD |      main |       16 |   PKCS7 |     602.1 ns |     3.56 ns |     3.16 ns |  1.00 |    0.00 | 0.0200 |     - |     - |     128 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
|      Decrypt_Cbc_ExactBufferSize | Job-SSHGXV |    branch |       16 |   PKCS7 |     564.5 ns |     5.03 ns |     4.70 ns |  0.94 |    0.01 | 0.0200 |     - |     - |     128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZZAIDD |      main |       16 |   PKCS7 |     598.6 ns |     4.96 ns |     4.64 ns |  1.00 |    0.00 | 0.0200 |     - |     - |     128 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
| Decrypt_Cbc_GuaranteedBufferSize | Job-SSHGXV |    branch |      128 |   PKCS7 |     593.6 ns |     4.50 ns |     4.21 ns |  1.00 |    0.01 | 0.0200 |     - |     - |     128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZZAIDD |      main |      128 |   PKCS7 |     590.7 ns |     6.75 ns |     6.32 ns |  1.00 |    0.00 | 0.0200 |     - |     - |     128 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
|      Decrypt_Cbc_ExactBufferSize | Job-SSHGXV |    branch |      128 |   PKCS7 |     602.4 ns |     4.33 ns |     3.84 ns |  1.00 |    0.01 | 0.0200 |     - |     - |     128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZZAIDD |      main |      128 |   PKCS7 |     602.2 ns |     3.65 ns |     3.23 ns |  1.00 |    0.00 | 0.0200 |     - |     - |     128 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
| Decrypt_Cbc_GuaranteedBufferSize | Job-SSHGXV |    branch |      129 |   PKCS7 |     594.3 ns |     7.26 ns |     6.79 ns |  1.00 |    0.02 | 0.0200 |     - |     - |     128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZZAIDD |      main |      129 |   PKCS7 |     593.8 ns |     9.23 ns |     8.63 ns |  1.00 |    0.00 | 0.0200 |     - |     - |     128 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
|      Decrypt_Cbc_ExactBufferSize | Job-SSHGXV |    branch |      129 |   PKCS7 |     609.0 ns |     3.75 ns |     3.50 ns |  1.00 |    0.01 | 0.0200 |     - |     - |     128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZZAIDD |      main |      129 |   PKCS7 |     608.1 ns |     2.92 ns |     2.59 ns |  1.00 |    0.00 | 0.0200 |     - |     - |     128 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
| Decrypt_Cbc_GuaranteedBufferSize | Job-SSHGXV |    branch |  5242880 |   PKCS7 | 460,490.4 ns |   952.68 ns |   795.53 ns |  0.74 |    0.01 |      - |     - |     - |     128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZZAIDD |      main |  5242880 |   PKCS7 | 617,425.1 ns | 7,254.54 ns | 6,785.90 ns |  1.00 |    0.00 |      - |     - |     - |     129 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
|      Decrypt_Cbc_ExactBufferSize | Job-SSHGXV |    branch |  5242880 |   PKCS7 | 476,455.1 ns | 2,884.06 ns | 2,556.64 ns |  0.78 |    0.00 |      - |     - |     - |     128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZZAIDD |      main |  5242880 |   PKCS7 | 606,969.7 ns | 1,131.37 ns | 1,002.93 ns |  1.00 |    0.00 |      - |     - |     - |     129 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
| Decrypt_Cbc_GuaranteedBufferSize | Job-SSHGXV |    branch |  5242881 |   PKCS7 | 462,148.7 ns | 2,065.94 ns | 1,831.40 ns |  0.76 |    0.00 |      - |     - |     - |     128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZZAIDD |      main |  5242881 |   PKCS7 | 605,905.5 ns | 1,016.27 ns |   848.63 ns |  1.00 |    0.00 |      - |     - |     - |     129 B |
|                                  |            |           |          |         |              |             |             |       |         |        |       |       |           |
|      Decrypt_Cbc_ExactBufferSize | Job-SSHGXV |    branch |  5242881 |   PKCS7 | 461,642.1 ns |   539.61 ns |   450.60 ns |  0.75 |    0.00 |      - |     - |     - |     128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZZAIDD |      main |  5242881 |   PKCS7 | 614,144.4 ns |   732.79 ns |   611.91 ns |  1.00 |    0.00 |      - |     - |     - |     129 B |
</details>