@vcsjones 2022-02-10
Okay. I think prior to this, we (I) need to lite-ify `AesCng` and `TripleDESCng`. Their one shot implementation is in `UniversalCryptoDecryptor` because they needed to be able to handle ncrypt keys, so it has a duplicate implementation of the one shot.

Since it sounds like we're open to "decrypt in to the user buffer" I will re-open a PR for that, but I would take to take care of the one-shot implementation on the public CNG types, first.